requirements.txt
beautifulsoup4==4.7.1
bunch==1.0.1
lxml==4.2.1
markdownify==0.4.1
matplotlib==2.1.1
nltk==3.4.1
numpy==1.13.3
pandas==0.24.2
pdfkit==0.6.1
pyminizip==0.2.4
PyPDF2==1.26.0
PyYAML==5.1
scikit_learn==0.20.3
scipy==1.2.1
